Refractor'd Components folder
Components folder makes more sense structurally now
