+++ ryders-rally/src/components/Home/_home.scss
@@ -1,0 +1,7 @@
+#title{
+  color: rgb(190, 146, 24) !important;
+  font-family: "Merienda", cursive  !important;
+  font-size: 40px;
+}
+
+
