+++ ryders-rally/src/components/Login/Login.jsx
@@ -2,26 +2,25 @@ import React, { useContext, useState } from 'react';
 import { Button, Form } from 'semantic-ui-react';
 import { useMutation } from '@apollo/react-hooks';
 import gql from 'graphql-tag';
+import "/_login.scss"
 
-import { AuthContext } from '../context/auth';
-import { useForm } from '../helper/hooks';
+import { AuthContext } from '../../context/auth';
+import { useForm } from '../../helper/hooks';
 
-function Register(props) {
+function Login(props) {
   const context = useContext(AuthContext);
   const [errors, setErrors] = useState({});
 
-  const { onChange, onSubmit, values } = useForm(registerUser, {
+  const { onChange, onSubmit, values } = useForm(loginUserCallback, {
     username: '',
-    email: '',
-    password: '',
-    confirmPassword: ''
+    password: ''
   });
 
-  const [addUser, { loading }] = useMutation(REGISTER_USER, {
+  const [loginUser, { loading }] = useMutation(LOGIN_USER, {
     update(
       _,
       {
-        data: { register: userData }
+        data: { login: userData }
       }
     ) {
       context.login(userData);
@@ -33,14 +32,14 @@ function Register(props) {
     variables: values
   });
 
-  function registerUser() {
-    addUser();
+  function loginUserCallback() {
+    loginUser();
   }
 
   return (
     <div className="form-container">
       <Form onSubmit={onSubmit} noValidate className={loading ? 'loading' : ''}>
-        <h1>Register</h1>
+        <h1>Login</h1>
         <Form.Input
           label="Username"
           placeholder="Username.."
@@ -48,15 +47,6 @@ function Register(props) {
           type="text"
           value={values.username}
           error={errors.username ? true : false}
-          onChange={onChange}
-        />
-        <Form.Input
-          label="Email"
-          placeholder="Email.."
-          name="email"
-          type="email"
-          value={values.email}
-          error={errors.email ? true : false}
           onChange={onChange}
         />
         <Form.Input
@@ -68,17 +58,8 @@ function Register(props) {
           error={errors.password ? true : false}
           onChange={onChange}
         />
-        <Form.Input
-          label="Confirm Password"
-          placeholder="Confirm Password.."
-          name="confirmPassword"
-          type="password"
-          value={values.confirmPassword}
-          error={errors.confirmPassword ? true : false}
-          onChange={onChange}
-        />
         <Button type="submit" primary>
-          Register
+          Login
         </Button>
       </Form>
       {Object.keys(errors).length > 0 && (
@@ -94,21 +75,9 @@ function Register(props) {
   );
 }
 
-const REGISTER_USER = gql`
-  mutation register(
-    $username: String!
-    $email: String!
-    $password: String!
-    $confirmPassword: String!
-  ) {
-    register(
-      registerInput: {
-        username: $username
-        email: $email
-        password: $password
-        confirmPassword: $confirmPassword
-      }
-    ) {
+const LOGIN_USER = gql`
+  mutation login($username: String!, $password: String!) {
+    login(username: $username, password: $password) {
       id
       email
       username
@@ -118,4 +87,4 @@ const REGISTER_USER = gql`
   }
 `;
 
-export default Register;
+export default Login;
