+++ ryders-rally/src/components/Register/Register.jsx
@@ -2,25 +2,26 @@ import React, { useContext, useState } from 'react';
 import { Button, Form } from 'semantic-ui-react';
 import { useMutation } from '@apollo/react-hooks';
 import gql from 'graphql-tag';
-import "../_login.scss"
 
-import { AuthContext } from '../context/auth';
-import { useForm } from '../helper/hooks';
+import { AuthContext } from '../../context/auth';
+import { useForm } from '../../helper/hooks';
 
-function Login(props) {
+function Register(props) {
   const context = useContext(AuthContext);
   const [errors, setErrors] = useState({});
 
-  const { onChange, onSubmit, values } = useForm(loginUserCallback, {
+  const { onChange, onSubmit, values } = useForm(registerUser, {
     username: '',
-    password: ''
+    email: '',
+    password: '',
+    confirmPassword: ''
   });
 
-  const [loginUser, { loading }] = useMutation(LOGIN_USER, {
+  const [addUser, { loading }] = useMutation(REGISTER_USER, {
     update(
       _,
       {
-        data: { login: userData }
+        data: { register: userData }
       }
     ) {
       context.login(userData);
@@ -32,14 +33,14 @@ function Login(props) {
     variables: values
   });
 
-  function loginUserCallback() {
-    loginUser();
+  function registerUser() {
+    addUser();
   }
 
   return (
     <div className="form-container">
       <Form onSubmit={onSubmit} noValidate className={loading ? 'loading' : ''}>
-        <h1>Login</h1>
+        <h1>Register</h1>
         <Form.Input
           label="Username"
           placeholder="Username.."
@@ -47,6 +48,15 @@ function Login(props) {
           type="text"
           value={values.username}
           error={errors.username ? true : false}
+          onChange={onChange}
+        />
+        <Form.Input
+          label="Email"
+          placeholder="Email.."
+          name="email"
+          type="email"
+          value={values.email}
+          error={errors.email ? true : false}
           onChange={onChange}
         />
         <Form.Input
@@ -58,8 +68,17 @@ function Login(props) {
           error={errors.password ? true : false}
           onChange={onChange}
         />
+        <Form.Input
+          label="Confirm Password"
+          placeholder="Confirm Password.."
+          name="confirmPassword"
+          type="password"
+          value={values.confirmPassword}
+          error={errors.confirmPassword ? true : false}
+          onChange={onChange}
+        />
         <Button type="submit" primary>
-          Login
+          Register
         </Button>
       </Form>
       {Object.keys(errors).length > 0 && (
@@ -75,9 +94,21 @@ function Login(props) {
   );
 }
 
-const LOGIN_USER = gql`
-  mutation login($username: String!, $password: String!) {
-    login(username: $username, password: $password) {
+const REGISTER_USER = gql`
+  mutation register(
+    $username: String!
+    $email: String!
+    $password: String!
+    $confirmPassword: String!
+  ) {
+    register(
+      registerInput: {
+        username: $username
+        email: $email
+        password: $password
+        confirmPassword: $confirmPassword
+      }
+    ) {
       id
       email
       username
@@ -87,4 +118,4 @@ const LOGIN_USER = gql`
   }
 `;
 
-export default Login;
+export default Register;
