+++ ryders-rally/src/components/Sharing/Sharing.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "../_sharing.scss";
+import "/_sharing.scss";
 
 
 export default function SharingBar() {
